Refactor backend to adhere to SOLID principles.
This commit refactors the backend to better follow SOLID principles, primarily the Single Responsibility Principle.

- A new `users` app has been created to separate user management from task management. All user-related views, serializers, and URLs have been moved to this new app.
- A `TaskService` has been introduced in the `tasks` app to encapsulate business logic, making the views thinner and the code more modular.
- The permission system has been refactored to be more extensible, with role-checking logic centralized in the `permissions.py` file.
- The project has been configured to correctly support the new app structure.
