+++ backend/apps/tasks/serializers.py
@@ -20,23 +20,3 @@ class TaskSerializer(serializers.ModelSerializer):
             if Task.objects.filter(title__iexact=value, created_by=user).exists():
                 raise serializers.ValidationError("You already have a task with this title.")
         return value
-
-
-class UserCreateSerializer(serializers.ModelSerializer):
-    password = serializers.CharField(write_only=True)
-
-    class Meta:
-        model = User
-        fields = ['id', 'username', 'email', 'password', 'first_name', 'last_name','is_staff','is_superuser']
-
-    def create(self, validated_data):
-        user = User.objects.create_user(
-            username=validated_data['username'],
-            email=validated_data.get('email', ''),
-            password=validated_data['password'],
-            first_name=validated_data.get('first_name', ''),
-            last_name=validated_data.get('last_name', ''),
-            is_staff = validated_data.get('is_staff',False),
-            is_superuser=validated_data.get('is_superuser',False),
-        )
-        return user
