+++ backend/core/settings/base.py
@@ -35,7 +35,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'apps.tasks'
+    'apps.tasks',
+    'apps.users',
 ]
 
 MIDDLEWARE = [
