+++ backend/core/urls.py
@@ -15,7 +15,10 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path,include
+from django.urls import path, include
+from drf_yasg import openapi
+from drf_yasg.views import get_schema_view
+from rest_framework import permissions
 from rest_framework_simplejwt.views import (
     TokenObtainPairView,
     TokenRefreshView, TokenBlacklistView,
@@ -24,12 +27,28 @@ from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 
+# Swagger config
+schema_view = get_schema_view(
+    openapi.Info(
+        title="Task Manager API",
+        default_version='v1',
+        description="Task management API with real-time updates, role-based access, and token auth.",
+        contact=openapi.Contact(email="you@example.com"),
+        license=openapi.License(name="MIT License"),
+    ),
+    public=True,
+    permission_classes=[permissions.AllowAny],
+)
+
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
     path('api/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
     path('api/logout', TokenBlacklistView.as_view(), name='token_blacklist'),
-    path('api/', include('apps.tasks.urls')),
+    path('api/tasks/', include('apps.tasks.urls')),
+    path('api/users/', include('apps.users.urls')),
+    path('swagger/', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
+    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
 ]
 
 # Serve static files in development
